add metrics runner to phenoxtractor.py
